Basic UDP push of colonists that have changed
For each colonist it finds, the mod will check if their hash has
changed. All changed colonists are added to a list that is then bundled
up and sent as a UDP packet.

The JSON functionality is a bit ugly right now since I manually
implemented it myself, but I'd rather get to working on the UDP server
and Frontend code than spend too long on JSON libraries right now.:w;: